@@ -61,38 +61,39 @@
     "@graphql-codegen/typescript-operations": "^2.4.2",
     "@graphql-eslint/eslint-plugin": "^3.10.4",
     "@jompx/jcdk": "file:../jompx/dist/packages/jcdk",
-    "@nrwl/angular": "^13.0.0",
-    "@nrwl/cli": "13.10.3",
-    "@nrwl/cypress": "13.10.5",
-    "@nrwl/eslint-plugin-nx": "13.10.5",
-    "@nrwl/jest": "13.10.5",
-    "@nrwl/linter": "13.10.5",
-    "@nrwl/nx-cloud": "13.3.1",
-    "@nrwl/workspace": "13.10.3",
+    "@nrwl/angular": "13.10.6",
+    "@nrwl/cli": "14.4.2",
+    "@nrwl/cypress": "14.4.2",
+    "@nrwl/eslint-plugin-nx": "14.4.2",
+    "@nrwl/jest": "14.4.2",
+    "@nrwl/linter": "14.4.2",
+    "@nrwl/nx-cloud": "14.2.0",
+    "@nrwl/workspace": "14.4.2",
     "@nxtend/capacitor": "^13.0.0",
     "@nxtend/ionic-angular": "^13.1.0",
     "@types/get-value": "^3.0.2",
-    "@types/jest": "^26.0.10",
+    "@types/jest": "27.4.1",
     "@types/node": "16.11.7",
     "@types/set-value": "^4.0.1",
-    "@typescript-eslint/eslint-plugin": "~5.18.0",
-    "@typescript-eslint/parser": "~5.18.0",
+    "@typescript-eslint/eslint-plugin": "5.29.0",
+    "@typescript-eslint/parser": "5.29.0",
     "aws-cdk": "2.31.1",
     "aws-organization-formation": "^0.9.20-beta.3",
     "cypress": "^9.1.0",
     "esbuild": "^0.14.27",
-    "eslint": "~8.12.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.10.3",
-    "jest": "^26.4.2",
+    "jest": "27.5.1",
     "jest-preset-angular": "11.1.1",
-    "nx": "13.10.3",
-    "prettier": "^2.5.1",
-    "ts-jest": "^26.2.0",
-    "ts-node": "^9.0.0",
-    "typescript": "~4.5.2"
+    "nx": "14.4.2",
+    "prettier": "2.6.2",
+    "ts-jest": "27.1.4",
+    "ts-node": "10.8.2",
+    "typescript": "4.7.4"
   },
   "peerDependencies": {
     "@jompx/datasource": "file:../jompx/dist/packages/datasource"
   }
 }
+
